+++ apps/react-3d/src/app/ui/nx-welcome.stories.tsx
@@ -4,9 +4,9 @@ import { NxWelcome } from './nx-welcome';
 import { within } from '@storybook/testing-library';
 import { expect } from '@storybook/jest';
 
-import withAPIRequest from './with-api-request';
+import withAPIRequest from '../with-api-request';
 import { DefaultBodyType, http, HttpResponse, PathParams } from 'msw';
-import { TanStackQueryMockData } from '../mocks/tan-stack-query-mock.data';
+import { TanStackQueryMockData } from '../../mocks/tan-stack-query-mock.data';
 
 const meta: Meta<typeof NxWelcome> = {
   component: NxWelcome,
